Feat: 버스킹 일정 삭제 시 크루장 인가 로직 추가 #S12P11E204-137

+++ backend/butter/src/main/java/com/ssafy/butter/domain/schedule/service/ScheduleServiceImpl.java
@@ -88,8 +88,13 @@ public class ScheduleServiceImpl implements ScheduleService {
     }
 
     @Override
-    public ScheduleResponseDTO deleteSchedule(Long id) {
+    public ScheduleResponseDTO deleteSchedule(AuthInfoDTO currentUser, Long id) {
+        Member member = memberService.findById(currentUser.id());
         Schedule schedule = scheduleRepository.findById(id).orElseThrow();
+        Crew crew = schedule.getCrew();
+        if (!crewMemberService.findByCrewAndMember(crew, member).getIsCrewAdmin()) {
+            throw new IllegalArgumentException("Current user is not crew admin");
+        }
         scheduleRepository.delete(schedule);
         return ScheduleResponseDTO.fromEntity(schedule);
     }
